+++ messages/el/base.php
@@ -1,9 +1,8 @@
 <?php
-
-return [
-    'Allows to start polls.' => '',
-    'At least one answer is required' => '',
-    'Cancel' => '',
-    'Polls' => '',
-    'Save' => 'Αποθήκευση',
-];
+return array (
+  'Allows to start polls.' => '',
+  'At least one answer is required' => '',
+  'Cancel' => 'Ακύρωση',
+  'Polls' => '',
+  'Save' => 'Αποθήκευση',
+);